Crate subspace_test_client

Subspace test client only.

Modules

auto_id_domain_chain_spec

Chain specification for the auto-id domain.

chain_spec

Chain specification for the test runtime.

evm_domain_chain_spec

Chain specification for the evm domain.

Functions

start_farmer

Run a farmer.

Type Aliases

Backend

The backend type being used by the test service.

Client

The client type being used by the test service.