Crate subspace_test_client

Subspace test client only.

Modules

chain_spec

Chain specification for the test runtime.

Structs

TestExecutorDispatch

Subspace native executor instance.

Functions

start_farmer

Run a farmer.

Type Definitions

Backend

The backend type being used by the test service.

Client

The client type being used by the test service.